set topic and description
